Restrict coordinators to storage pods
This makes it possible to write PDBs that safely permit individual node
draining even when a single pod from each of multiple storage classes
are present on that node. See #507 and #316 for more context.

Possibly this should automatically fall back to using all stateful
storage classes in very small clusters? Or perhaps a configuration
switch?

Signed-off-by: Robert Collins <robert.collins@cognite.com>