Recreating pods that are not excluded.
This is a bugfix for a situation that can happen when a cluster is in the process of being scaled down at the same time as some nodes are down. If one or more of the nodes that are down are marked for removal but weren't fully drained before they went down, we need to bring them back up again so that they can be drained before actually being shut down.

This fixes #970.

Bug fix (non-breaking change which fixes an issue).

This implementation of the fix relies on a previous run of the `removeProcessGroups` subreconciler having called `SetExcluded()` on process groups that FDB has confirmed to be fully excluded from the cluster (and saved that status to Kubernetes).

- Unit tests (including new tests in controllers/add_pods_test.go)
- Taking down some storage nodes in a dev FDB cluster and then initiating a scale-down, and seeing that the storage nodes are brought back up again and drained before they are excluded

*Did you update relevant documentation within this repository?*

TODO

*If this change is adding new functionality, do we need to describe it in our user manual?*

N/A.

*If this change is adding or removing subreconcilers, have we updated the core technical design doc to reflect that?*

N/A.

*If this change is adding new safety checks or new potential failure modes, have we documented and how to debug potential issues?*

*Are there any follow-up issues that we should pursue in the future?*

TODO

*Does this introduce new defaults that we should re-evaluate in the future?*

No.