Recreating pods that are not excluded. (#8)
This is a bugfix for a situation that can happen when a cluster is in the process of being scaled down at the same time as some nodes are down. If one or more of the nodes that are down are marked for removal but weren't fully drained before they went down, we need to bring them back up again so that they can be drained before actually being shut down.

This implementation of the fix relies on a previous run of the `removeProcessGroups` subreconciler having called `SetExcluded()` on process groups that FDB has confirmed to be fully excluded from the cluster (and saved that status to Kubernetes).

This fixes https://github.com/FoundationDB/fdb-kubernetes-operator/issues/970.